controllers/user: Move `PUT /user/:user_id` endpoints to a dedicated module
This makes the code slightly easier to find, since these endpoints are unrelated to the ones in the `/me` namespace.
