+++ src/controllers/user.rs
@@ -1,3 +1,6 @@
 pub mod me;
 pub mod other;
 pub mod session;
+pub mod update;
+
+pub use update::update_user;
