+++ src/controllers/user/me.rs
@@ -5,20 +5,16 @@ use axum::Json;
 use diesel::prelude::*;
 use diesel_async::async_connection_wrapper::AsyncConnectionWrapper;
 use http::request::Parts;
-use lettre::Address;
-use secrecy::{ExposeSecret, SecretString};
 use serde_json::Value;
 use std::collections::HashMap;
 
 use crate::app::AppState;
 use crate::controllers::helpers::pagination::{Paginated, PaginationOptions};
 use crate::controllers::helpers::{ok_true, Paginate};
-use crate::models::{
-    CrateOwner, Email, Follow, NewEmail, OwnerKind, User, Version, VersionOwnerAction,
-};
+use crate::models::{CrateOwner, Follow, OwnerKind, User, Version, VersionOwnerAction};
 use crate::schema::{crate_owners, crates, emails, follows, users, versions};
 use crate::tasks::spawn_blocking;
-use crate::util::errors::{bad_request, server_error, AppResult};
+use crate::util::errors::{bad_request, AppResult};
 use crate::util::BytesRequest;
 use crate::views::{EncodableMe, EncodablePrivateUser, EncodableVersion, OwnedCrate};
 
@@ -110,85 +106,6 @@ pub async fn updates(app: AppState, req: Parts) -> AppResult<Json<Value>> {
     .await
 }
 
-/// Handles the `PUT /users/:user_id` route.
-pub async fn update_user(
-    state: AppState,
-    Path(param_user_id): Path<i32>,
-    req: BytesRequest,
-) -> AppResult<Response> {
-    let conn = state.db_write().await?;
-    spawn_blocking(move || {
-        let conn: &mut AsyncConnectionWrapper<_> = &mut conn.into();
-
-        use self::emails::user_id;
-        use diesel::insert_into;
-
-        let auth = AuthCheck::default().check(&req, conn)?;
-        let user = auth.user();
-
-        // need to check if current user matches user to be updated
-        if user.id != param_user_id {
-            return Err(bad_request("current user does not match requested user"));
-        }
-
-        #[derive(Deserialize)]
-        struct UserUpdate {
-            user: User,
-        }
-
-        #[derive(Deserialize)]
-        struct User {
-            email: Option<String>,
-        }
-
-        let user_update: UserUpdate =
-            serde_json::from_slice(req.body()).map_err(|_| bad_request("invalid json request"))?;
-
-        let user_email = match &user_update.user.email {
-            Some(email) => email.trim(),
-            None => return Err(bad_request("empty email rejected")),
-        };
-
-        if user_email.is_empty() {
-            return Err(bad_request("empty email rejected"));
-        }
-
-        user_email
-            .parse::<Address>()
-            .map_err(|_| bad_request("invalid email address"))?;
-
-        let new_email = NewEmail {
-            user_id: user.id,
-            email: user_email,
-        };
-
-        let token = insert_into(emails::table)
-            .values(&new_email)
-            .on_conflict(user_id)
-            .do_update()
-            .set(&new_email)
-            .returning(emails::token)
-            .get_result(conn)
-            .map(SecretString::new)
-            .map_err(|_| server_error("Error in creating token"))?;
-
-        // This swallows any errors that occur while attempting to send the email. Some users have
-        // an invalid email set in their GitHub profile, and we should let them sign in even though
-        // we're trying to silently use their invalid address during signup and can't send them an
-        // email. They'll then have to provide a valid email address.
-        let email = UserConfirmEmail {
-            user_name: &user.gh_login,
-            domain: &state.emails.domain,
-            token,
-        };
-
-        let _ = state.emails.send(user_email, email);
-
-        ok_true()
-    })
-    .await
-}
-
 /// Handles the `PUT /confirm/:email_token` route
 pub async fn confirm_user_email(state: AppState, Path(token): Path<String>) -> AppResult<Response> {
     let conn = state.db_write().await?;
@@ -204,48 +121,6 @@ pub async fn confirm_user_email(state: AppState, Path(token): Path<String>) -> A
         if updated_rows == 0 {
             return Err(bad_request("Email belonging to token not found."));
         }
-
-        ok_true()
-    })
-    .await
-}
-
-/// Handles `PUT /user/:user_id/resend` route
-pub async fn regenerate_token_and_send(
-    state: AppState,
-    Path(param_user_id): Path<i32>,
-    req: Parts,
-) -> AppResult<Response> {
-    let conn = state.db_write().await?;
-    spawn_blocking(move || {
-        let conn: &mut AsyncConnectionWrapper<_> = &mut conn.into();
-
-        use diesel::dsl::sql;
-        use diesel::update;
-
-        let auth = AuthCheck::default().check(&req, conn)?;
-        let user = auth.user();
-
-        // need to check if current user matches user to be updated
-        if user.id != param_user_id {
-            return Err(bad_request("current user does not match requested user"));
-        }
-
-        conn.transaction(|conn| -> AppResult<_> {
-            let email: Email = update(Email::belonging_to(user))
-                .set(emails::token.eq(sql("DEFAULT")))
-                .get_result(conn)
-                .optional()?
-                .ok_or_else(|| bad_request("Email could not be found"))?;
-
-            let email1 = UserConfirmEmail {
-                user_name: &user.gh_login,
-                domain: &state.emails.domain,
-                token: email.token,
-            };
-
-            state.emails.send(&email.email, email1).map_err(Into::into)
-        })?;
 
         ok_true()
     })
@@ -315,31 +190,4 @@ pub async fn update_email_notifications(app: AppState, req: BytesRequest) -> App
         ok_true()
     })
     .await
-}
-
-pub struct UserConfirmEmail<'a> {
-    pub user_name: &'a str,
-    pub domain: &'a str,
-    pub token: SecretString,
-}
-
-impl crate::email::Email for UserConfirmEmail<'_> {
-    fn subject(&self) -> String {
-        "crates.io: Please confirm your email address".into()
-    }
-
-    fn body(&self) -> String {
-        // Create a URL with token string as path to send to user
-        // If user clicks on path, look email/user up in database,
-        // make sure tokens match
-
-        format!(
-            "Hello {user_name}! Welcome to crates.io. Please click the
-link below to verify your email address. Thank you!\n
-https://{domain}/confirm/{token}",
-            user_name = self.user_name,
-            domain = self.domain,
-            token = self.token.expose_secret(),
-        )
-    }
 }
